fix: snowflake ID alignment — default_option_ids i32→i64, attribute_option ID gen, label_template source_id
- Change default_option_ids from Vec<i32> to Vec<i64> across entire stack
  (shared models, cloud DB/API, edge repository/handlers, Tauri commands)
- Fix attribute_option cloud ops: generate snowflake_id when assigned_id is None
  instead of relying on SQLite autoincrement
- Fix label_template Console CRUD: use source_id AS id pattern consistent with
  all other resources, update/delete by source_id instead of PG BIGSERIAL id
- Fix data_transfer handler: remove i32↔i64 casts in default_option_ids remap

+++ crab-cloud/src/api/store/attribute.rs
@@ -244,7 +244,7 @@ pub struct BindAttributeRequest {
     pub is_required: bool,
     #[serde(default)]
     pub display_order: i32,
-    pub default_option_ids: Option<Vec<i32>>,
+    pub default_option_ids: Option<Vec<i64>>,
 }
 
 #[derive(serde::Deserialize)]
